--- packages/verite/lib/builders/credential-fulfillment-builder.ts
@@ -1,9 +1,0 @@
-import { DecodedCredentialFulfillment } from "../../types"
-
-export class CredentialFulfillmentBuilder {
-  _builder: Partial<DecodedCredentialFulfillment>
-
-  constructor() {
-    this._builder = {}
-  }
-}
