+++ packages/verite/lib/builders/index.ts
@@ -1,5 +1,8 @@
 export * from "./common"
 export * from "./credential-application-builder"
+export * from "./credential-application-wrapper-builder"
+export * from "./credential-response-builder"
+export * from "./credential-response-wrapper-builder"
 export * from "./credential-payload-builder"
 export * from "./input-descriptors"
 export * from "./credential-manifest-builder"
